Use Turn instead of int for turn variable, add Deck class, separate tests
This allows us to have a str representation of "T1" not just "1" and is useful
in model variable names/solution output/just generally. It's a little weak in
that Turn(1) + Turn(2) yields an int not a Turn but it's good enough.

Bundle constraints along with deck size and add some helper functions by using
a Deck class.

Tests are now in a separate file and run with pytest. This gives more useful
output when a test fails.